+++ src/getStepStartStates.ts
@@ -204,6 +204,7 @@ export function getResourceManagerSteps(): GetApiSteps {
       SecuritySteps.SECURITY_CENTER_CONTACTS,
       SecuritySteps.PRICING_CONFIGURATIONS,
       MonitorSteps.MONITOR_LOG_PROFILES,
+      MonitorSteps.MONITOR_ACTIVITY_LOG_ALERTS,
     ],
     executeLastSteps: [
       AdvisorSteps.RECOMMENDATIONS,