+++ src/steps/resource-manager/databases/sql/index.test.ts
@@ -14,7 +14,7 @@ import {
 } from '.';
 import { entities, relationships } from './constants';
 import { configFromEnv } from '../../../../../test/integrationInstanceConfig';
-import { getMockAccountEntity } from '../../../../../test/helpers/getMockAccountEntity';
+import { getMockAccountEntity } from '../../../../../test/helpers/getMockEntity';
 import {
   separateDiagnosticSettingsEntities,
   separateDiagnosticSettingsRelationships,