+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added `azure_monitor_activity_log_alert` entities
+
 ## 5.19.0 - 2021-03-26
 
 ### Added